Replace run queue with a simple linked list
The fiber control block has a link field 'next' for linking
waiters, use that for the run queue as well. A linked list
implementation of a priority queue should suffice when there
are only a handful of fibers (which is assumed to be the case
on a microcontroller). This new implementation also lets us
support 256 priority levels.

The job of _schedule() is now to simply grab the next item
in the run queue linked list, so write it in assembly and
bake it into yield() to lose the call overhead.

Replace wake() with a re-entrant implementation written in
assembly. This enables use of wake() in re-entrant interrupt
handlers.