Fix wake() replacing currently running fiber
So, obviously it doesn't work to have the current FCB be
the head of the run queue, since a higher-priority fiber
may replace it in the queue while it's still executing.
Make a separate run queue and modify yield() and wake()
as necessary to fix that embarrassing oversight.

While at it, namespace the privately used symbols to recude
the chances of a collision, and tidy up the public interface
function argument naming, too.